MTN: Refactored debug_firehose to separate file
